refactor: change file structure in routers folder
