+++ src/routers/controllers/index.ts
@@ -1,0 +1,2 @@
+export * as userController from './user.conroller';
+export * as groupCrontroller from './group.conroller';
